backend/requirements.txt
aiosqlite==0.22.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
bcrypt==3.2.2
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.123.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.6.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.1.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
ruff==0.1.6
pre-commit==3.5.0
google-auth==2.23.4
facebook-sdk==3.1.0
PyJWT==2.8.0
